Change arc weight type from int to float for precise thresholds
All arc classes now accept float weights instead of int:
- Arc.__init__(weight: float = 1)
- Arc.set_weight(weight: float)
- InhibitorArc.__init__(weight: float = 1)
- TestArc.__init__(weight: float = 1)
- CurvedArc.__init__(weight: float = 1)
- CurvedInhibitorArc.__init__(weight: float = 1)

This allows precise threshold values like:
- Inhibitor arc weight = 2.5 mM (not just 2 or 3)
- Stoichiometric coefficients = 1.5 (e.g., fractional reactions)
- Test arc thresholds = 0.5 (activation at half concentration)

Example 08 already uses float weights (2.5, 2.0) in JSON,
now properly supported by type system.

+++ src/shypn/netobjs/curved_arc.py
@@ -23,7 +23,7 @@ class CurvedArc(Arc):
     # Curve offset as percentage of line length
     CURVE_OFFSET_RATIO = 0.20  # 20% of line length
     
-    def __init__(self, source, target, id: int, name: str, weight: int = 1):
+    def __init__(self, source, target, id: int, name: str, weight: float = 1):
         """Initialize a curved arc.
         
         Args:
